Change --concurrency-factor from int multiplier to float64 CPU ratio
On large machines (e.g. 96 CPUs), the old integer multiplier on
runtime.NumCPU() created excessive workers with no way to reduce below
1× CPU count. Changing to a float64 ratio (default 0.5) lets users
dial concurrency down on large hosts (e.g. 0.5 on 96 CPUs = 48 workers).

- CLI flag: IntFlag → Float64Flag, default 0.5, valid range (0.0, 4.0]
- All struct fields and config: int → float64
- Computation: math.Round(NumCPU * factor) with floor of 1
- API handler, jobs config, HTML reporter updated accordingly
- Documentation and OpenAPI spec updated

+++ internal/consistency/diff/repset_diff.go
@@ -47,7 +47,7 @@ type RepsetDiffCmd struct {
 	clusterNodes      []map[string]any
 	database          types.Database
 	ConnectionPool    *pgxpool.Pool
-	ConcurrencyFactor int
+	ConcurrencyFactor float64
 	BlockSize         int
 	CompareUnitSize   int
 	Output            string
